🧹 Split out components
